--- alembic/versions/5c762ec28bd2_initial_migration.py
@@ -1,111 +1,0 @@
-"""Initial migration
-
-Revision ID: 5c762ec28bd2
-Revises: 
-Create Date: 2025-01-15 09:06:43.008791
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '5c762ec28bd2'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('blocked_token',
-    sa.Column('token_id', sa.String(length=255), nullable=False),
-    sa.Column('expired_at', sa.DateTime(), nullable=False),
-    sa.PrimaryKeyConstraint('token_id')
-    )
-    op.create_table('user',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('username', sa.String(length=20), nullable=True),
-    sa.Column('nickname', sa.String(length=20), nullable=True),
-    sa.Column('email', sa.String(length=100), nullable=False),
-    sa.Column('password', sa.String(length=20), nullable=False),
-    sa.Column('address', sa.String(length=100), nullable=True),
-    sa.Column('phone_number', sa.String(length=20), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_user_email'), 'user', ['email'], unique=True)
-    op.create_index(op.f('ix_user_nickname'), 'user', ['nickname'], unique=True)
-    op.create_index(op.f('ix_user_username'), 'user', ['username'], unique=True)
-    op.create_table('blog',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('blog_name', sa.String(length=100), nullable=False),
-    sa.Column('address_name', sa.String(length=100), nullable=False),
-    sa.Column('description', sa.String(length=255), nullable=True),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('main_image_url', sa.String(length=255), nullable=True),
-    sa.Column('user_id', sa.BigInteger(), nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('user_id', name='unique_user_blog')
-    )
-    op.create_table('category',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('name', sa.String(length=50), nullable=False),
-    sa.Column('level', sa.Integer(), nullable=False),
-    sa.Column('blog_id', sa.BigInteger(), nullable=False),
-    sa.Column('parent_id', sa.BigInteger(), nullable=True),
-    sa.CheckConstraint('level IN (1, 2)', name='valid_category_level'),
-    sa.ForeignKeyConstraint(['blog_id'], ['blog.id'], ),
-    sa.ForeignKeyConstraint(['parent_id'], ['category.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('name', 'blog_id', name='unique_category_per_blog')
-    )
-    op.create_index(op.f('ix_category_name'), 'category', ['name'], unique=False)
-    op.create_table('article',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('title', sa.String(length=20), nullable=False),
-    sa.Column('content', sa.Text(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('blog_id', sa.BigInteger(), nullable=False),
-    sa.Column('category_id', sa.BigInteger(), nullable=False),
-    sa.ForeignKeyConstraint(['blog_id'], ['blog.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['category_id'], ['category.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_article_title'), 'article', ['title'], unique=False)
-    op.create_table('comment',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('content', sa.String(length=500), nullable=False),
-    sa.Column('level', sa.Integer(), nullable=False),
-    sa.Column('secret', sa.Integer(), nullable=False),
-    sa.Column('user_id', sa.BigInteger(), nullable=True),
-    sa.Column('article_id', sa.BigInteger(), nullable=True),
-    sa.Column('parent_id', sa.BigInteger(), nullable=True),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
-    sa.ForeignKeyConstraint(['article_id'], ['article.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['parent_id'], ['comment.id'], ),
-    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('comment')
-    op.drop_index(op.f('ix_article_title'), table_name='article')
-    op.drop_table('article')
-    op.drop_index(op.f('ix_category_name'), table_name='category')
-    op.drop_table('category')
-    op.drop_table('blog')
-    op.drop_index(op.f('ix_user_username'), table_name='user')
-    op.drop_index(op.f('ix_user_nickname'), table_name='user')
-    op.drop_index(op.f('ix_user_email'), table_name='user')
-    op.drop_table('user')
-    op.drop_table('blocked_token')
-    # ### end Alembic commands ###
