--- alembic/versions/7440806423ba_initial_migration.py
@@ -1,30 +1,0 @@
-"""Initial migration
-
-Revision ID: 7440806423ba
-Revises: 5c762ec28bd2
-Create Date: 2025-01-15 09:10:44.807431
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '7440806423ba'
-down_revision: Union[str, None] = '5c762ec28bd2'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
