+++ alembic/versions/ecac98e10032_initial_migration.py
@@ -1,0 +1,45 @@
+"""Initial migration
+
+Revision ID: ecac98e10032
+Revises: 
+Create Date: 2025-01-15 10:33:58.720882
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'ecac98e10032'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('comment',
+    sa.Column('id', sa.BigInteger(), nullable=False),
+    sa.Column('content', sa.String(length=500), nullable=False),
+    sa.Column('level', sa.Integer(), nullable=False),
+    sa.Column('secret', sa.Integer(), nullable=False),
+    sa.Column('user_name', sa.String(length=20), nullable=False),
+    sa.Column('user_id', sa.BigInteger(), nullable=True),
+    sa.Column('article_id', sa.BigInteger(), nullable=True),
+    sa.Column('parent_id', sa.BigInteger(), nullable=True),
+    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
+    sa.ForeignKeyConstraint(['article_id'], ['article.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['parent_id'], ['comment.id'], ),
+    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('comment')
+    # ### end Alembic commands ###
