+++ pyproject.toml
@@ -20,9 +20,8 @@ aiomysql = "^0.2.0"
 pyopenssl = "^24.3.0"
 authlib = "^1.4.0"
 httpx = "^0.28.1"
-jwt = "^1.3.1"
-pyjwt = "^2.10.1"
 asyncmy = "^0.2.10"
+pyjwt = "^2.10.1"
 
 
 [tool.poetry.group.dev.dependencies]
